+++ src/main/kotlin/io/red/financesK/transaction/model/InstallmentInfo.kt
@@ -1,9 +1,7 @@
 package io.red.financesK.transaction.model
 
-import java.math.BigDecimal
-
 data class InstallmentInfo(
     val totalInstallments: Int? = null,
     val currentInstallment: Int? = null,
-    val installmentValue: BigDecimal? = null,
+    val installmentValue: Int? = null,
 )
